Changed some field data types to int64 To fix crashes on 32-bit systems. Also, it explicitly says so at https://core.telegram.org/bots/api

+++ models/document.go
@@ -7,5 +7,5 @@ type Document struct {
 	Thumb        *PhotoSize `json:"thumb,omitempty"`
 	FileName     string     `json:"file_name,omitempty"`
 	MimeType     string     `json:"mime_type,omitempty"`
-	FileSize     int        `json:"file_size,omitempty"`
+	FileSize     int64      `json:"file_size,omitempty"`
 }
